requirement.txt
fastapi ~= 0.89.1
python-jose ~= 3.3.0
pymysql ~= 1.0.2
sqlalchemy ~= 1.4.46
uvicorn ~= 0.20.0